+++ src/mem.h
@@ -10,5 +10,5 @@
 
 #include <cstdint>
 
-void kmpSearch(const void *full, int fullLen, const void *sub, int subLen, void(*func)(int, void *), void *userp);
-void md5Hash(const void *buf, int len, uint8_t output[16]);
+void kmpSearch(const void *full, size_t fullLen, const void *sub, size_t subLen, void(*func)(int, void *), void *userp);
+void md5Hash(const void *buf, size_t len, uint8_t output[16]);
